requirements.txt
pyyaml
pythorhead @ git+https://github.com/db0/pythorhead
transformers
torchvision
torch